Seed overhaul: split bootstrap vs demo, auto-seed empty DBs
