+++ app/templates/admin/data_upload/expense_accounts.html
@@ -127,7 +127,7 @@
         <td><strong>Default Spend Type</strong></td>
         <td>No</td>
         <td><code>spend_type</code>, <code>default_spend_type</code></td>
-        <td>Code or name (e.g., DIVVY, BANK, HOTEL_FEE)</td>
+        <td>Code or name (e.g., DIVVY, BANK)</td>
       </tr>
       <tr>
         <td><strong>Spend Type Mode</strong></td>
